+++ .eslintrc.json
@@ -13,6 +13,7 @@
   "rules": {
     "linebreak-style": ["error", "windows"],
     "no-shadow": "off",
+    "import/prefer-default-export": 0,
     "no-param-reassign": "off",
     "eol-last": "off",
     "import/extensions": [ 1, {
