--- tests/addItem.test.js
@@ -1,5 +1,0 @@
-import addItem from './addItem.js';
-
-test('Properly adding a new item', () => {
-  expect(addItem()).toBe(1);
-});
