--- tests/deleteItem.test.js
@@ -1,5 +1,0 @@
-import deleteAllCheckedItems from './deleteItem.js';
-
-test('Properly adding a new item', () => {
-  expect(deleteAllCheckedItems()).toBe(1);
-});
